refactor: replace commons/ with layered modules
`commons/` had become a bag for anything shared: a pure tokenisation algorithm, an
IndexedDB schema, DOM mutation, the messaging protocol and the i18next setup all
lived side by side, so its name said where the code was imported from rather than
what it does. `utils.ts` was the worst of it -- extension settings, the kanji
filter database and a Tailwind class merger in one file, which meant most modules
pulled in `idb` and `tailwind-merge` to reach a settings getter.

Split it by what the code depends on:

  core/       toKanjiToken, tokenize      -- npm packages only
  constants/  cn                          -- npm packages only
  storage/    settings, kanjiFilterDB     -- constants, assets
  messaging/  message                     -- constants, core
  dom/        addFurigana, selectElement  -- constants, storage, messaging
  i18n                                    -- constants, storage, assets

Dependencies now run strictly one way, with no cycles: `core` is reachable from
everything and reaches nothing, which is what makes it unit-testable without the
extension APIs, and `entrypoints` sits on top as the only consumer of all of it.

Pure moves and import rewrites; no behaviour change. Verified with typecheck,
lint, the unit suite, a production build, and the Playwright suite. The e2e run
is unchanged from the same run on the parent commit: the three "Kanji filter
page" failures reproduce identically before and after, and
"Automatically add furigana when lang is ja" is flaky in both directions.
