+++ extension/__TEST__/e2e/popup.spec.ts
@@ -1,5 +1,5 @@
 import type { Page } from "@playwright/test";
-import { ExtStorage, type GeneralSettings } from "@/commons/constants";
+import { ExtStorage, type GeneralSettings } from "@/constants";
 import { describe, expect, test } from "../fixtures";
 
 describe("Extension popup page", () => {
