+++ extension/src/constants.ts
