+++ extension/src/core/toKanjiToken.ts
