+++ extension/src/dom/selectElement.ts
