+++ extension/src/entrypoints/background/listeners/onMarkActiveMessage.ts
@@ -1,4 +1,4 @@
-import { type ExtensionMessage, onMessage } from "@/commons/message";
+import { type ExtensionMessage, onMessage } from "@/messaging/message";
 
 type MarkStyle = {
   fillStyle: string;
