+++ extension/src/entrypoints/options/routes/KanjiFilter/components/ImportKanjiFilterButton.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { isKanji, isKatakana } from "wanakana";
 import { z } from "zod";
-import type { FilterRule } from "@/commons/constants";
+import type { FilterRule } from "@/constants";
 import { PopupTransition } from "@/entrypoints/options/components/PopupTransition";
 import { useKanjiFiltersStore } from "../store";
