+++ extension/src/entrypoints/options/routes/KanjiFilter/components/KanjiFilterDashboard.tsx
@@ -2,7 +2,8 @@ import { Dialog, DialogPanel, DialogTitle } from "@headlessui/react";
 import { saveAs } from "file-saver";
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
-import { cn, DB, getKanjiFilterDB } from "@/commons/utils";
+import { cn } from "@/cn";
+import { DB, getKanjiFilterDB } from "@/storage/kanjiFilterDB";
 import { PopupTransition } from "../../../components/PopupTransition";
 import { useKanjiFiltersStore } from "../store";
 import { BatchCreateKanjiFilterButton } from "./BatchCreateKanjiFilterButton";
