+++ extension/src/entrypoints/options/routes/Selector/components/SelectorDashboard.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { useTranslation } from "react-i18next";
 
 import defaultSelectorRules from "@/assets/rules/selector.json";
-import { cn } from "@/commons/utils";
+import { cn } from "@/cn";
 import { PopupTransition } from "../../../components/PopupTransition";
 import { useSelectorsStore } from "../store";
 import { CreateNewSelectorRuleButton } from "./CreateNewSelectorRuleButton";
