+++ extension/src/entrypoints/popup/components/Button.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/commons/utils";
+import { cn } from "@/cn";
 import { ToolTip } from "./ToolTip";
 
 interface ButtonProps {
