+++ extension/src/entrypoints/popup/components/ColorPicker.tsx
@@ -3,7 +3,7 @@ import { Popover, PopoverButton, PopoverPanel, Transition } from "@headlessui/re
 import { clamp } from "es-toolkit";
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
-import { cn } from "@/commons/utils";
+import { cn } from "@/cn";
 
 interface ColorPickerProps {
   color: string;
