+++ extension/src/entrypoints/popup/root.tsx
@@ -11,9 +11,9 @@ import HiraganaIcon from "@/assets/icons/Hiragana.svg?react";
 import PowerIcon from "@/assets/icons/Power.svg?react";
 import SettingIcon from "@/assets/icons/Setting.svg?react";
 import ShareIcon from "@/assets/icons/Share.svg?react";
-import { DisplayMode, ExtStorage, FuriganaType, SelectMode } from "@/commons/constants";
-import { sendMessage } from "@/commons/message";
-import { cn } from "@/commons/utils";
+import { cn } from "@/cn";
+import { DisplayMode, ExtStorage, FuriganaType, SelectMode } from "@/constants";
+import { sendMessage } from "@/messaging/message";
 import { Button } from "./components/Button";
 import { CheckBox } from "./components/CheckBox";
 import { ColorPicker } from "./components/ColorPicker";
