+++ extension/src/storage/settings.ts
@@ -1,20 +1,13 @@
-import { type ClassValue, clsx } from "clsx";
-import { type DBSchema, openDB } from "idb";
-import { twMerge } from "tailwind-merge";
-
-import defaultKanjiFilterRules from "@/assets/rules/filter.json";
 import defaultSelectorRules from "@/assets/rules/selector.json";
-
 import {
   DisplayMode,
   ExtStorage,
-  type FilterRule,
   FuriganaType,
   type GeneralSettings,
   type MoreSettings,
   SelectMode,
   type SelectorRule,
-} from "./constants";
+} from "@/constants";
 
 export const generalSettingsFallback = {
   [ExtStorage.AutoMode]: true,
@@ -84,37 +77,3 @@ export const customSelectors = storage.defineItem<SelectorRule[]>("local:customR
   version: 1,
   fallback: defaultSelectorRules,
 });
-
-export const DB = {
-  name: "kanjiFilterDB",
-  version: 1,
-  onlyTable: "kanjiFilterTable",
-} as const;
-interface KanjiFilterDB extends DBSchema {
-  [DB.onlyTable]: {
-    key: string;
-    value: FilterRule;
-  };
-}
-
-export const getKanjiFilterDB = async () => {
-  const db = await openDB<KanjiFilterDB>(DB.name, DB.version, {
-    /**
-     * @param transaction
-     * Don't use `db.transaction(...)`, the upgrade callback will run a version change transaction,
-     * and new transactions can't be created until this transaction ends.
-     */
-    upgrade(db, _, __, transaction) {
-      db.createObjectStore(DB.onlyTable, { keyPath: "kanji" });
-      const store = transaction.objectStore(DB.onlyTable);
-      for (const rule of defaultKanjiFilterRules) {
-        store.put(rule);
-      }
-    },
-  });
-  return db;
-};
-
-export function cn(...args: ClassValue[]) {
-  return twMerge(clsx(args));
-}
